requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
requests
langchain-openai
chromadb
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain-groq
langserve
fastapi
uvicorn
sse_starlette
duckduckgo-search
langchainhub
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
cassio 
datasets 
langchain 
openai 
tiktoken
PyPDF2
langchain-google-genai
google-generativeai
langchain-astradb
gradio
langchain_nvidia_ai_endpoints
crewai
crewai_tools
yt_dlp==2023.11.14
youtube-transcript-api==0.6.1